chore: move src/features/cli to src/cli
