+++ packages/dynamic-instructions/src/cli/commands/generate-client-types/generate-client-types-from-file.ts
