+++ packages/dynamic-instructions/src/cli/commands/generate-client-types/register-command.ts
