+++ packages/dynamic-instructions/src/cli/commands/index.ts
