+++ packages/dynamic-instructions/src/cli/index.ts
