+++ packages/dynamic-instructions/src/cli/program.ts
